requirements.txt
langchain
langchain-openai
python-dotenv
langchain-google-genai
numpy
requests
streamlit
langchain-community